validate post

+++ PRH_PostService/Application/Commands/Posts/DeletePost/DeletePostCommandHandler.cs
@@ -1,12 +1,15 @@
 ﻿using Application.Commons;
+using Application.Commons.Tools;
 using Application.Interfaces.Repository;
 using MediatR;
+using Microsoft.AspNetCore.Http;
 using System.Net;
 
 
 namespace Application.Commands.Posts.DeletePost
 {
-    public class DeletePostCommandHandler(IPostRepository postRepository) : IRequestHandler<DeletePostCommand, BaseResponse<string>>
+    public class DeletePostCommandHandler(IPostRepository postRepository, IHttpContextAccessor httpContextAccessor) 
+        : IRequestHandler<DeletePostCommand, BaseResponse<string>>
     {
         public async Task<BaseResponse<string>> Handle(DeletePostCommand request, CancellationToken cancellationToken)
         {
@@ -18,6 +21,32 @@ namespace Application.Commands.Posts.DeletePost
             };
             try
             {
+                var userId = Authentication.GetUserIdFromHttpContext(httpContextAccessor.HttpContext);
+                if (userId == null)
+                {
+                    response.Success = false;
+                    response.Message = "Không có quyền để truy cập";
+                    response.StatusCode = (int)HttpStatusCode.Unauthorized;
+                    return response;
+                }
+
+                var post = await postRepository.GetByIdAsync(request.Id);
+                if (post == null)
+                {
+                    response.Success = false;
+                    response.Message = "Bài viết không tồn tại";
+                    response.StatusCode = (int)HttpStatusCode.NotFound;
+                    return response;
+                }
+
+                if (post.UserId != userId)
+                {
+                    response.Success = false;
+                    response.Message = "Bạn không có quyền xoá bài viết này";
+                    response.StatusCode = (int)HttpStatusCode.Forbidden;
+                    return response;
+                }
+
                 await postRepository.DeleteAsync(request.Id);
                 response.StatusCode = (int)HttpStatusCode.OK;
                 response.Success = true;
